refactor: extract-class 'RotaryLoop' into shared/
This is the stoppable, sleeping, async-friendly periodic polling loop.
